refactor(#3479): extract turn_bridge streaming-edit-text + watcher-orphan-cleanup to submodules
#3479 Phase-1 rank-4 behavior-preserving extraction. Pure move, zero logic
change. Shrinks the frozen giant turn_bridge/mod.rs from 6702 -> 6535 prod LoC.

Cluster A -> turn_bridge/streaming_edit_text.rs (88 prod):
  build_turn_bridge_streaming_edit_text, bridge_pre_submission_tui_prompt_error,
  bridge_tui_transport_error_should_skip_quiescence + their tests.

Cluster B -> turn_bridge/watcher_orphan_cleanup.rs (119 prod):
  should_delete_bridge_created_watcher_orphan_response,
  should_retry_watcher_orphan_spinner_cleanup,
  record_watcher_orphan_spinner_cleanup,
  spawn_watcher_orphan_spinner_cleanup_retry + their tests.

Both modules re-exported from mod.rs so all call sites stay byte-identical.
WatcherHandoffClaimOutcome widened private -> pub(super) (visibility only) so the
child module can reference it; the enum stays in mod.rs (still used there). The
terminal_delivery_response_after_offset test stays in mod.rs (unrelated target).
Function bodies are byte-identical; one signature reflowed by rustfmt due to the
pub(in ...) visibility prefix width. Lowered the giant baseline in the toml,
registered both child modules in change-surfaces.md, regenerated ARCHITECTURE.md.
