+++ src/services/discord/turn_bridge/mod.rs
@@ -11,12 +11,14 @@ mod single_message_footer;
 mod skill_usage;
 mod stale_resume;
 mod status_panel;
+mod streaming_edit_text;
 mod terminal_controller_cutover;
 mod terminal_delivery;
 mod tmux_runtime;
 mod turn_analytics;
 mod voice_completion;
 mod watcher_handoff;
+mod watcher_orphan_cleanup;
 
 use super::gateway::TurnGateway;
 use super::restart_report::{RestartCompletionReport, clear_restart_report, save_restart_report};
@@ -61,6 +63,10 @@ pub(super) use stale_resume::result_event_has_stale_resume_error;
 pub(in crate::services::discord) use status_panel::{
     complete_status_panel_v2_with_http, normalize_status_panel_message_id,
 };
+pub(super) use streaming_edit_text::{
+    bridge_pre_submission_tui_prompt_error, bridge_tui_transport_error_should_skip_quiescence,
+    build_turn_bridge_streaming_edit_text,
+};
 pub(crate) use tmux_runtime::TmuxCleanupPolicy;
 pub(super) use tmux_runtime::bind_cancel_token_tmux_runtime;
 pub(super) use tmux_runtime::cancel_active_token;
@@ -68,6 +74,10 @@ pub(super) use tmux_runtime::cancel_token_has_tmux_session;
 pub(super) use tmux_runtime::handoff_interrupted_message;
 pub(super) use tmux_runtime::stale_inflight_message;
 pub(super) use tmux_runtime::stop_active_turn;
+pub(super) use watcher_orphan_cleanup::{
+    record_watcher_orphan_spinner_cleanup, should_delete_bridge_created_watcher_orphan_response,
+    should_retry_watcher_orphan_spinner_cleanup, spawn_watcher_orphan_spinner_cleanup_retry,
+};
 
 /// #2452 H6 graduation: schedule the history-aware auto-retry via the
 /// gateway's `_with_completion` variant, then release the
@@ -908,172 +918,15 @@ mod sentinel_overwrite_clamp_tests {
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum WatcherHandoffClaimOutcome {
+pub(super) enum WatcherHandoffClaimOutcome {
     None,
     ReusedExisting,
     Spawned,
 }
 
-fn should_delete_bridge_created_watcher_orphan_response(
-    status_panel_v2_enabled: bool,
-    watcher_handoff_claim_outcome: WatcherHandoffClaimOutcome,
-    bridge_created_response_placeholder_msg_id: Option<MessageId>,
-    current_msg_id: MessageId,
-) -> bool {
-    status_panel_v2_enabled
-        && watcher_handoff_claim_outcome == WatcherHandoffClaimOutcome::ReusedExisting
-        && bridge_created_response_placeholder_msg_id == Some(current_msg_id)
-}
-
-fn should_retry_watcher_orphan_spinner_cleanup(
-    outcome: &super::placeholder_cleanup::PlaceholderCleanupOutcome,
-) -> bool {
-    matches!(
-        outcome,
-        super::placeholder_cleanup::PlaceholderCleanupOutcome::Failed {
-            class: super::placeholder_cleanup::PlaceholderCleanupFailureClass::LifecycleFailure,
-            ..
-        }
-    )
-}
-
-fn record_watcher_orphan_spinner_cleanup(
-    shared: &SharedData,
-    provider: &ProviderKind,
-    channel_id: ChannelId,
-    message_id: MessageId,
-    tmux_session_name: Option<&str>,
-    outcome: super::placeholder_cleanup::PlaceholderCleanupOutcome,
-    source: &'static str,
-) {
-    if let super::placeholder_cleanup::PlaceholderCleanupOutcome::Failed { class, detail } =
-        &outcome
-    {
-        let ts = chrono::Local::now().format("%H:%M:%S");
-        tracing::warn!(
-            "  [{ts}] ⚠ watcher orphan spinner cleanup failed ({}) for channel {} msg {}: {}",
-            class.as_str(),
-            channel_id.get(),
-            message_id.get(),
-            detail
-        );
-    }
-    shared
-        .ui
-        .placeholder_cleanup
-        .record(super::placeholder_cleanup::PlaceholderCleanupRecord {
-            provider: provider.clone(),
-            channel_id,
-            message_id,
-            tmux_session_name: tmux_session_name.map(str::to_string),
-            operation: super::placeholder_cleanup::PlaceholderCleanupOperation::DeleteNonterminal,
-            outcome,
-            source,
-        });
-}
-
-fn spawn_watcher_orphan_spinner_cleanup_retry(
-    shared: Arc<SharedData>,
-    provider: ProviderKind,
-    gateway: Arc<dyn TurnGateway>,
-    channel_id: ChannelId,
-    message_id: MessageId,
-    tmux_session_name: Option<String>,
-) {
-    const RETRY_DELAYS: &[std::time::Duration] = &[
-        std::time::Duration::from_secs(2),
-        std::time::Duration::from_secs(5),
-        std::time::Duration::from_secs(15),
-    ];
-
-    super::task_supervisor::spawn_observed("watcher_orphan_spinner_cleanup_retry", async move {
-        for (attempt, delay) in RETRY_DELAYS.iter().enumerate() {
-            tokio::time::sleep(*delay).await;
-            let outcome = match gateway.delete_message(channel_id, message_id).await {
-                Ok(()) => super::placeholder_cleanup::PlaceholderCleanupOutcome::Succeeded,
-                Err(error) => super::placeholder_cleanup::classify_delete_error(&error),
-            };
-            let committed = outcome.is_committed();
-            let should_retry = should_retry_watcher_orphan_spinner_cleanup(&outcome);
-            record_watcher_orphan_spinner_cleanup(
-                shared.as_ref(),
-                &provider,
-                channel_id,
-                message_id,
-                tmux_session_name.as_deref(),
-                outcome,
-                "turn_bridge_watcher_orphan_spinner_cleanup_retry",
-            );
-            if committed || !should_retry {
-                return;
-            }
-            if attempt + 1 == RETRY_DELAYS.len() {
-                let ts = chrono::Local::now().format("%H:%M:%S");
-                tracing::warn!(
-                    "  [{ts}] ⚠ watcher orphan spinner cleanup exhausted retries for channel {} msg {}",
-                    channel_id.get(),
-                    message_id.get()
-                );
-            }
-        }
-    });
-}
-
-fn build_turn_bridge_streaming_edit_text(
-    status_panel_v2_enabled: bool,
-    current_portion: &str,
-    status_block: &str,
-    provider: &ProviderKind,
-) -> String {
-    if status_panel_v2_enabled {
-        super::formatting::build_status_panel_streaming_edit_text(
-            current_portion,
-            status_block,
-            provider,
-        )
-    } else {
-        super::formatting::build_streaming_placeholder_text(current_portion, status_block)
-    }
-}
-
 #[cfg(test)]
 mod streaming_edit_text_tests {
     use super::*;
-
-    #[test]
-    fn status_panel_v2_streaming_edit_moves_processing_footer_to_response_message() {
-        let rendered = build_turn_bridge_streaming_edit_text(
-            true,
-            "E2E-CODEX-1-OK\n- Working on the backend now",
-            "⠙ 계속 처리 중",
-            &ProviderKind::Codex,
-        );
-
-        assert_eq!(
-            rendered,
-            "E2E-CODEX-1-OK\n- Working on the backend now\n\n⠙ 계속 처리 중"
-        );
-    }
-
-    #[test]
-    fn legacy_streaming_edit_keeps_processing_footer() {
-        let rendered = build_turn_bridge_streaming_edit_text(
-            false,
-            "Partial answer",
-            "⠙ 계속 처리 중",
-            &ProviderKind::Codex,
-        );
-
-        assert_eq!(rendered, "Partial answer\n\n⠙ 계속 처리 중");
-    }
-
-    #[test]
-    fn status_panel_v2_empty_streaming_edit_keeps_placeholder() {
-        let rendered =
-            build_turn_bridge_streaming_edit_text(true, "", "⠙ 계속 처리 중", &ProviderKind::Codex);
-
-        assert_eq!(rendered, "⠙ 계속 처리 중");
-    }
 
     #[test]
     fn empty_response_notice_is_delivery_only_not_history_payload() {
@@ -1083,178 +936,6 @@ mod streaming_edit_text_tests {
 
         assert_eq!(rendered, "(No response)");
         assert!(full_response.is_empty());
-    }
-
-    #[test]
-    fn watcher_reuse_deletes_only_bridge_created_response_placeholder() {
-        let bridge_placeholder = MessageId::new(20);
-
-        assert!(should_delete_bridge_created_watcher_orphan_response(
-            true,
-            WatcherHandoffClaimOutcome::ReusedExisting,
-            Some(bridge_placeholder),
-            bridge_placeholder,
-        ));
-        assert!(!should_delete_bridge_created_watcher_orphan_response(
-            true,
-            WatcherHandoffClaimOutcome::Spawned,
-            Some(bridge_placeholder),
-            bridge_placeholder,
-        ));
-        assert!(!should_delete_bridge_created_watcher_orphan_response(
-            true,
-            WatcherHandoffClaimOutcome::ReusedExisting,
-            Some(bridge_placeholder),
-            MessageId::new(10),
-        ));
-        assert!(!should_delete_bridge_created_watcher_orphan_response(
-            false,
-            WatcherHandoffClaimOutcome::ReusedExisting,
-            Some(bridge_placeholder),
-            bridge_placeholder,
-        ));
-        assert!(!should_delete_bridge_created_watcher_orphan_response(
-            true,
-            WatcherHandoffClaimOutcome::None,
-            Some(bridge_placeholder),
-            bridge_placeholder,
-        ));
-    }
-
-    #[test]
-    fn watcher_orphan_spinner_cleanup_retries_only_lifecycle_failures() {
-        use super::super::placeholder_cleanup::{
-            PlaceholderCleanupFailureClass, PlaceholderCleanupOutcome,
-        };
-
-        assert!(should_retry_watcher_orphan_spinner_cleanup(
-            &PlaceholderCleanupOutcome::Failed {
-                class: PlaceholderCleanupFailureClass::LifecycleFailure,
-                detail: "HTTP 500".to_string(),
-            }
-        ));
-        assert!(!should_retry_watcher_orphan_spinner_cleanup(
-            &PlaceholderCleanupOutcome::Failed {
-                class: PlaceholderCleanupFailureClass::PermissionOrRoutingDiagnostic,
-                detail: "HTTP 403".to_string(),
-            }
-        ));
-        assert!(!should_retry_watcher_orphan_spinner_cleanup(
-            &PlaceholderCleanupOutcome::Succeeded
-        ));
-        assert!(!should_retry_watcher_orphan_spinner_cleanup(
-            &PlaceholderCleanupOutcome::AlreadyGone
-        ));
-    }
-}
-
-fn bridge_pre_submission_tui_prompt_error(provider: &ProviderKind, full_response: &str) -> bool {
-    let Some(error_text) = full_response
-        .trim_start()
-        .strip_prefix("Error:")
-        .map(str::trim_start)
-    else {
-        return false;
-    };
-    match provider {
-        ProviderKind::Claude => {
-            crate::services::claude_tui::input::is_prompt_ready_timeout_error(error_text)
-        }
-        ProviderKind::Codex => {
-            crate::services::codex_tui::input::is_prompt_ready_timeout_error(error_text)
-        }
-        _ => false,
-    }
-}
-
-fn bridge_tui_transport_error_should_skip_quiescence(
-    provider: &ProviderKind,
-    runtime_kind: Option<crate::services::agent_protocol::RuntimeHandoffKind>,
-    full_response: &str,
-) -> bool {
-    let Some(error_text) = full_response
-        .trim_start()
-        .strip_prefix("Error:")
-        .map(str::trim_start)
-    else {
-        return false;
-    };
-
-    match (provider, runtime_kind) {
-        (
-            ProviderKind::Claude,
-            Some(crate::services::agent_protocol::RuntimeHandoffKind::ClaudeTui),
-        ) => {
-            bridge_pre_submission_tui_prompt_error(provider, full_response)
-                || error_text == "Timeout waiting for output file"
-                || error_text.starts_with("timeout waiting for claude tui transcript file")
-                || error_text.contains("claude tui session died")
-        }
-        (
-            ProviderKind::Codex,
-            Some(crate::services::agent_protocol::RuntimeHandoffKind::CodexTui),
-        ) => {
-            bridge_pre_submission_tui_prompt_error(provider, full_response)
-                || error_text == "Timeout waiting for output file"
-                || error_text.contains("codex tui session died")
-        }
-        _ => false,
-    }
-}
-
-#[cfg(test)]
-mod pre_submission_tui_prompt_error_tests {
-    use super::*;
-
-    #[test]
-    fn classifier_matches_wrapped_readiness_errors() {
-        assert!(bridge_pre_submission_tui_prompt_error(
-            &ProviderKind::Claude,
-            "Error: timeout waiting for claude tui follow-up prompt input readiness after 45s; reason=prompt_marker_not_detected; previous_tui_turn_still_running=true; capture_available=true",
-        ));
-        assert!(bridge_pre_submission_tui_prompt_error(
-            &ProviderKind::Codex,
-            "Error: timeout waiting for codex tui follow-up prompt input readiness after 45s; reason=composer_not_detected; previous_tui_turn_still_running=true; capture_available=true",
-        ));
-        assert!(!bridge_pre_submission_tui_prompt_error(
-            &ProviderKind::Claude,
-            "Error: claude tui session died during follow-up output reading",
-        ));
-        assert!(!bridge_pre_submission_tui_prompt_error(
-            &ProviderKind::Claude,
-            "timeout waiting for claude tui follow-up prompt input readiness after 45s",
-        ));
-    }
-
-    #[test]
-    fn tui_transport_errors_skip_quiescence_only_for_matching_tui_runtime() {
-        use crate::services::agent_protocol::RuntimeHandoffKind;
-
-        assert!(bridge_tui_transport_error_should_skip_quiescence(
-            &ProviderKind::Claude,
-            Some(RuntimeHandoffKind::ClaudeTui),
-            "Error: Timeout waiting for output file",
-        ));
-        assert!(bridge_tui_transport_error_should_skip_quiescence(
-            &ProviderKind::Claude,
-            Some(RuntimeHandoffKind::ClaudeTui),
-            "Error: timeout waiting for claude tui transcript file after 120s; capture_available=true; prompt_marker_detected=true; prompt_draft_detected=false",
-        ));
-        assert!(bridge_tui_transport_error_should_skip_quiescence(
-            &ProviderKind::Codex,
-            Some(RuntimeHandoffKind::CodexTui),
-            "Error: timeout waiting for codex tui follow-up prompt input readiness after 45s; reason=composer_not_detected; previous_tui_turn_still_running=true; capture_available=true",
-        ));
-        assert!(!bridge_tui_transport_error_should_skip_quiescence(
-            &ProviderKind::Claude,
-            Some(RuntimeHandoffKind::LegacyTmuxWrapper),
-            "Error: Timeout waiting for output file",
-        ));
-        assert!(!bridge_tui_transport_error_should_skip_quiescence(
-            &ProviderKind::Claude,
-            Some(RuntimeHandoffKind::ClaudeTui),
-            "Error: upstream API returned 500",
-        ));
     }
 }
